Removes unused includes

+++ custommayaconstructs/tools/voxelpaintcontextcommand.h
@@ -36,11 +36,11 @@ public:
         MArgParser ap = parser();
         if (ap.isFlagSet("-r")) {
             double v; ap.getFlagArgument("-r", 0, v);
-            fCtx->setSelectRadius(v);
+            fCtx->setSelectRadius(static_cast<float>(v));
         }
         if (ap.isFlagSet("-v")) {
             double v; ap.getFlagArgument("-v", 0, v);
-            fCtx->setSelectStrength(v);
+            fCtx->setSelectStrength(static_cast<float>(v));
         }
         if (ap.isFlagSet("-ifs")) {
             int v; ap.getFlagArgument("-ifs", 0, v);
